LayoutTests/imported/w3c/web-platform-tests/custom-elements/parser/parser-fallsback-to-unknown-element-expected.txt

CONSOLE MESSAGE: TypeError: The result of constructing a custom element must be a HTMLElement
CONSOLE MESSAGE: TypeError: The result of constructing a custom element must be a HTMLElement
CONSOLE MESSAGE: line 35: ReferenceError: Cannot access uninitialized variable.
CONSOLE MESSAGE: line 41: Bad
CONSOLE MESSAGE: line 1321: Error: Tried to create a test with file_is_test

PASS Custom Elements: Changes to the HTML parser 
